+++ src/frontend/src/views/index.ts
@@ -5,7 +5,6 @@ export * from "./HomeView";
 export * from "./LlmPromptView";
 export * from "./AgentCanvasView";
 export * from "./AgentMarketplaceView";
-export * from "./WeatherView";
 export * from "./LandingPage";
 export * from "./LoginView";
 export * from "./RegisterView";